server/requirements.txt
# Jarvis Backend Server Requirements

Flask>=3.0.0
flask-cors>=4.0.0
groq>=0.4.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
